requirements.txt
requests
Jinja2
pandas
python-dotenv
